requirements.txt
yfinance==0.2.40
requests==2.31.0
azure-eventhub==5.11.5
pyspark==3.5.0
azure-cosmos==4.7.0
azure-storage-blob==12.19.1
apache-airflow==2.8.1
pyyaml==6.0.1
